Extract code from loopObjAutocomplete into isWithinLoop file
